Logic bug fix
Previously rewrote a file every run due to a logic error. The intent was only to write a file when it did not exist in a specific place. Script was looking in the wrong place for the file and wrote it every time. Now looks in correct place for file.